fmu-v6x: increase UART5 buffer size
The same fix had to be done for the fmu-v5x:
https://github.com/PX4/PX4-Autopilot/pull/14932

Prior to this mavlink FTP was not working at all
and the mavlink stream error rate was quite high.
With this fix the error rate is still not zero, but
mavlink FTP works and AuterionOS can update the
configuration successfully.