uavcan: also check sd-root/ext_autostart for fw
AuterionOS can inject config files such as mixer definitions
or binaries. One limitation of our injection process though is
that all injected files end up in the directory

/fs/microsd/temp

which is then renamed to

/fs/microsd/new

only if AuterionOS successfully copied all files. Then PX4 is
rebooted, and PX4 renames

/fs/microsd/new -> /fs/microsd/ext_autostart

This is to avoid modification of files while they are in use.
And it allows to roll-back an incomplete update.

The downside is, that the AuterionOS injection process does not
allow writing files directly to the root of the SD "/fs/microsd/"
where uavcan expects to find the firmware binaries. The solution
provided by this commit is to extend the uavcan module to check
/fs/microsd/ext_autostart/uavcan for uavcan binaries and then
copy them to /fs/microsd/ufw for the uavcan library to find them
and apply the update.